pyproject.toml
[project]
name = "linkedin-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "linkedin-api>=2.3.1",
    "mcp[cli]>=1.27.1",
    "python-dotenv>=1.2.2",
]
